Move contracts and tests out of subdirectories
