+++ contracts/GatedSingleEditionMintable.sol
