+++ contracts/ISingleEditionMintable.sol
