+++ contracts/ISingleEditionMintableCreator.sol
